lib: Compiler warning fixes for 32bit systems

+++ src/lib/test-istream-tee.c
@@ -30,7 +30,7 @@ static void test_istream_tee_tailing(const char *str)
 	for (len = 1; len < TEST_BUF_SIZE; len += delta) {
 		test_istream_set_size(test_input, len);
 		for (i = 0; i < CHILD_COUNT; i++) {
-			test_assert_idx(i_stream_read(child_input[i]) == delta, len);
+			test_assert_idx(i_stream_read(child_input[i]) == (int)delta, len);
 			test_assert_idx(!tee_i_stream_child_is_waiting(child_input[i]), len);
 			test_assert_idx(i_stream_read(child_input[i]) == 0, len);
 			test_assert_idx(!tee_i_stream_child_is_waiting(child_input[i]), len);
@@ -42,7 +42,7 @@ static void test_istream_tee_tailing(const char *str)
 
 	test_istream_set_size(test_input, len);
 	for (i = 0; i < CHILD_COUNT; i++) {
-		test_assert(i_stream_read(child_input[i]) == delta);
+		test_assert(i_stream_read(child_input[i]) == (int)delta);
 		test_assert(i_stream_read(child_input[i]) == -2);
 		test_assert(!tee_i_stream_child_is_waiting(child_input[i]));
 	}
@@ -64,7 +64,7 @@ static void test_istream_tee_tailing(const char *str)
 		}
 		i_stream_skip(child_input[lagger], delta);
 		for (i = 0; i < CHILD_COUNT; i++) {
-			test_assert(i_stream_read(child_input[i]) == delta);
+			test_assert(i_stream_read(child_input[i]) == (int)delta);
 			test_assert(i_stream_read(child_input[i]) == -2);
 			test_assert(!tee_i_stream_child_is_waiting(child_input[i]));
 		}
